Replace service call with hardcoded data responses
