--- VideoApi/VideoApi.Services/Clients/SupplierStub/Models/Room.cs
@@ -1,7 +1,0 @@
-namespace VideoApi.Services.Clients.SupplierStub.Models;
-
-public class Room
-{
-    [JsonPropertyName("label")]
-    public string Label { get; set; }
-}
