--- VideoApi/VideoApi.Services/Clients/SupplierStub/Requests/AddRoomRequest.cs
@@ -1,10 +1,0 @@
-using System.Collections.Generic;
-using VideoApi.Services.Clients.SupplierStub.Models;
-
-namespace VideoApi.Services.Clients.SupplierStub.Requests;
-
-public class AddRoomRequest
-{
-    [JsonPropertyName("rooms")]
-    public List<Room> Rooms { get; set; }
-}
