+++ charts/vh-video-api/values.yaml
@@ -85,4 +85,3 @@ java:
     SERVICES__CALLBACKURI: https://vh-video-web.{{ .Values.global.environment }}.platform.hmcts.net/callback
     UseStub: false
     SERVICES__VIDEOAPIURL: https://vh-video-api.{{ .Values.global.environment }}.platform.hmcts.net/
-    SUPPLIERSTUBCONFIGURATION__APIURL: https://vh-supplier-stub-api.com
